LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A1_T2-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: Parse error
S12.7_A1_T2

PASS Expected parsing failure

TEST COMPLETE
